### 2020-01-07 #2000 @whedon
[PRE REVIEW]: mhealthtools: A Modular R Package for Extracting Features from Mobile and Wearable Sensor Data
**Submitting author:** @philerooski (<a href="http://orcid.org/0000-0002-6967-2185">Phillip Snyder</a>)
**Repository:** <a href="https://github.com/Sage-Bionetworks/mhealthtools" target ="_blank">https://github.com/Sage-Bionetworks/mhealthtools</a>
**Version:** 0.2.0
