### 2020-01-23 #2000 @trallard
@philerooski let's start by finding some reviewers. Have you got someone in mind already? if not you can check the list of registered reviewers :<https://bit.ly/joss-reviewers> and suggest some
(⚠️ please make sure to not use the @ when adding their names to this issue)
