### 2020-01-23 #2000 @philerooski
@trallard There is a paper similar to this that has been published in JOSS (https://joss.theoj.org/papers/10.21105/joss.01778) that was reviewed by Abhishektha Boppana (`abhishektha`), although this reviewer is not familiar with R. Arne Küderle (`AKuederle`) is familiar with the conceptual aspects of this paper (IMUs, motion/gait analysis), but they do not list R as a programming language that they are familiar with, either. There are many reviewers familiar with both R and bioinformatics: `Ebedthan`, `kanekalla`, `dcassol`, ...
