### 2020-02-03 #2000 @AKuederle
Yes, I would be willing to, but because of my limited R knowledge, I might not be able to comment on aspects of code quality and correctness (?). If this is not an issue, I would be happy to perform the review.
