### 2020-02-17 Sage-Bionetworks/mhealthtools#206 @Ebedthan
Add clear guidelines for support seekers
Hello Team,

For easing the task of peoples who could run into trouble and seek support using your package, could you add a section showing how to find such support on this repos?

Thanks.
