### 2020-02-17 #2106 @Ebedthan
Hello @philerooski,
Just opened an [issue](https://github.com/Sage-Bionetworks/mhealthtools/issues/206) on the mhealthtools repository for adding details for support seekers as of community guidelines.

Regards.
