### 2020-02-17 Sage-Bionetworks/mhealthtools#207 @Ebedthan
Fix references inappropriately cited
Hi,

As another issue, you should fix the below inappropriately cited references:
- Citation 1:
The mPower study,**parkinson disease** mobile data collected using **researchkit**. should be The mPower study, **Parkinson disease** mobile data collected using **ResearchKit**.
- Citation 2:
**NPJ** Digital Medicine should be **npj** Digital medicine
- Citation 4:
**Gent, P. van, Bruin, J. de,** Kris, & Fernandes, G. should be **van Gent, P., de Bruin, J.**, Kris, & Fernandes, G.
- Citation 6 and 7:
Same as Citation 2
- Citation 8 and 9:
Please add the issue or volume number and the DOI of the paper.
- Citation 10: The title
New sensor
and wearable technologies to aid in the diagnosis and treatment monitoring of **parkinsons disease** should be New sensor and wearable technologies to aid in the diagnosis and treatment monitoring of **Parkinson's disease**.
- Citation 11:
Same as Citation 2 and add DOI.
- Citation 12:
If there is no clear information on this paper (year of publication, source, DOI, etc.), please find another one that supports your affirmation in the text.
- Citation 13:
Change parkinson to Parkinson and add DOI.
- Citation 14:
Change parkinson to Parkinson.

Thanks.
