+++ paper/paper.bib
@@ -1,7 +1,8 @@
 @article{bot_2016,
-  title={The mPower study, Parkinson disease mobile data collected using ResearchKit},
+  title={The mPower study, {Parkinson} disease mobile data collected using ResearchKit},
   url={https://www.nature.com/articles/sdata201611},
-  journal={Nature News},
+  DOI={10.1038/sdata.2016.11},
+  journal={Nature Scientific Data},
   publisher={Nature Publishing Group},
   author={Bot, Brian M. and Suver, Christine and Neto, Elias Chaibub and Kellen, Michael and Klein, Arno and Bare, Christopher and Doerr, Megan and Pratap, Abhishek and Wilbanks, John and Dorsey, E. Ray and Friend, Stephen H.},
   year={2016},
@@ -14,7 +15,7 @@
   url={https://www.nature.com/articles/s41746-019-0090-4},
   DOI={10.1038/s41746-019-0090-4},
   number={1},
-  journal={NPJ Digital Medicine},
+  journal={npj Digital Medicine},
   author={Coravos, Andrea and Khozin, Sean and Mandl, Kenneth D.},
   year={2019},
   month={Mar}
@@ -62,7 +63,7 @@
   url={https://www.nature.com/articles/s41746-019-0078-0},
   DOI={10.1038/s41746-019-0078-0},
   number={1},
-  journal={NPJ Digital Medicine},
+  journal={npj Digital Medicine},
   author={Jacobson, Nicholas C. and Weingarden, Hilary and Wilhelm, Sabine},
   year={2019},
   month={Feb}
@@ -74,7 +75,7 @@
   url={https://www.nature.com/articles/s41746-019-0084-2},
   DOI={10.1038/s41746-019-0084-2},
   number={1},
-  journal={NPJ Digital Medicine},
+  journal={npj Digital Medicine},
   author={Kourtis, Lampros C. and Regele, Oliver B. and Wright, Justin M. and Jones, Graham B.},
   year={2019},
   month={Feb}
@@ -82,7 +83,10 @@
 
 @article{kumar_nilsen_2013,
   title={Mobile health technology evaluation: the mHealth evidence workshop},
+  volume={45},
+  number={2},
   url={https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3803146/},
+  DOI={10.1016/j.amepre.2013.03.017},
   journal={American Journal of Preventive Medicine},
   publisher={U.S. National Library of Medicine},
   author={Kumar, Santosh and Nilsen, Wendy J and Abernethy, Amy and Atienza, Audie and Patrick, Kevin and Pavel, Misha and Riley, William T and Shar, Albert and Spring, Bonnie and Spruijt-Metz, Donna and et al.},
@@ -91,17 +95,20 @@
 }
 
 @article{majumder_deen_2019,
-         title={Smartphone Sensors for Health Monitoring and Diagnosis},
-         url={https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6539461/},
-         journal={Sensors (Basel, Switzerland)},
-         publisher={MDPI},
-         author={Majumder, Sumit and Deen, M Jamal},
-         year={2019},
-         month={May}
+  title={Smartphone Sensors for Health Monitoring and Diagnosis},
+  volume={19},
+  number={9},
+  url={https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6539461/},
+  DOI={10.3390/s19092164},
+  journal={Sensors (Basel, Switzerland)},
+  publisher={MDPI},
+  author={Majumder, Sumit and Deen, M Jamal},
+  year={2019},
+  month={May}
 }
 
 @article{monje_2019,
-  title={New Sensor and Wearable Technologies to Aid in the Diagnosis and Treatment Monitoring of Parkinsons Disease},
[…]
