+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{bot_2016,
-  title={The mPower study, {Parkinson} disease mobile data collected using ResearchKit},
+  title={The mPower study, {Parkinson} disease mobile data collected using {ResearchKit}},
   url={https://www.nature.com/articles/sdata201611},
   DOI={10.1038/sdata.2016.11},
   journal={Nature Scientific Data},
@@ -70,7 +70,7 @@
 }
 
 @article{kourtis_2019,
-  title={Digital biomarkers for Alzheimer’s disease: the mobile/wearable devices opportunity},
+  title={Digital biomarkers for {Alzheimer’s} disease: the mobile/wearable devices opportunity},
   volume={2},
   url={https://www.nature.com/articles/s41746-019-0084-2},
   DOI={10.1038/s41746-019-0084-2},
@@ -82,7 +82,7 @@
 }
 
 @article{kumar_nilsen_2013,
-  title={Mobile health technology evaluation: the mHealth evidence workshop},
+  title={Mobile health technology evaluation: the {mHealth} evidence workshop},
   volume={45},
   number={2},
   url={https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3803146/},
