### 2020-02-20 Sage-Bionetworks/mhealthtools#207 @philerooski
I've made all changes in #208 except one:

> Citation 12:
If there is no clear information on this paper (year of publication, source, DOI, etc.), please find another one that supports your affirmation in the text.

This may not be possible. The text it supports ("The heart rate features [generated by mhealthtools] have been used as part of a validation study") is extremely specific to this tool, which, being a brand new software library, has only been used in studies internal to my organization. The cited paper is currently being drafted.

If it's acceptable to keep citation 12 as is I can merge in #208 and close this issue.
