### 2020-02-21 Sage-Bionetworks/mhealthtools#209 @AKuederle
[Joss Review] Package
Hey everybody,

I will just use this issue to collect some thought about your package. Feel free to add your comments and opinions and we can figure out if there are things that would need to be addressed in a separate focused issue.

Major issues (things I think you be addressed):
- [ ] Provide more information about the vignettes in your README. I think they will be the starting point for most people who want to get started with the package.
- [ ] To build the vignettes `knitr` is required. The installation fails otherwise. I know most people have `knitr` installed, but this should be mentioned somewhere or listed in the dependencies.

Suggestions (things I thing would be cool to address):

- [ ] A link to to the devtools package in the installation section would be awesome to make things really easy for newer R users.
- [ ] Your Linux issue guide is specific for apt based systems (or maybe only ubuntu). This is fine but should be mentioned. Maybe a sentence like: "For `apt` based systems (e.g., Debian or Ubuntu) you can run the commands below. For other distros (or in case these commands should fail), please check the package repository of your distro for the correct command and package names to use.
- [ ] Contributing to a repo does not just mean adding code. Your contribution guide should also mention where people should open issues and how to contribute new ideas/suggestions
- [ ] The file `data.R` contains comments about the specific number of rows for datatype. Doesn't this depend on my input data? Or am I misunderstanding something.
- [ ] I am not sure if this is typical in the R-world, but a hosted documentation with browsable function doc-strings would be awesome (thinking ReadTheDocs in the Python world)

Awesome things:
- Including a Docker version is awesome for testing
- Test for all mature functionality
- Test data included
- CI automation for tests
- Overall repo structure really clean
- Documentation of single functions is clear and complete
- Long form documentaion is awesome

General comments:
- The described Linux installation issues were not present in Manjaro Linux
- Unfortunally, I can not comment on the overall quality of your code, as I am not familiar enough with R
