### 2020-02-21 Sage-Bionetworks/mhealthtools#210 @AKuederle
[Joss Review] Paper
Hey everybody,

I will collect my thoughts on the paper here (I will try not to repeat points form the other reviewers)

- Paper is well written and nicely explains the structure of the model

Issues:
- [x] You mention that various other libraries exist. I would like to see a comparison of these packages to your module to further strengthen your statement of need.
