### 2020-02-26 7b59006
add note to unpublished manuscript citation
