+++ paper/paper.bib
@@ -134,6 +134,8 @@
 @unpublished{omberg_2020,
   author = {Omberg, Larsson},
   title = "Validation of a Smartphone Measure of VO2max",
+  note = {mhealthtools heart rate features were used as part of this study which,
+          as of February 2020, is currently under draft within Sage Bionetworks.}
 }
 
 @article{perumal_2018,
