### 2020-02-26 Sage-Bionetworks/mhealthtools#207 @philerooski
Hi @Ebedthan, yes -- I added a `note` tag to the citation which now makes it a valid `@unpublished` citation according to https://www.openoffice.org/bibliographic/bibtex-defs.html
