### 2020-02-26 #2106 @Ebedthan
My review is complete. Recommend accept it with changes provided by authors. This package will ease stuff on extracting features from sensor.
