### 2020-02-26 Sage-Bionetworks/mhealthtools#209 @philerooski
Some really good suggestions here @AKuederle. I'll provide comments on some points and address the others in one or more pull requests.

> To build the vignettes knitr is required. The installation fails otherwise. I know most people have knitr installed, but this should be mentioned somewhere or listed in the dependencies.

R package dependencies are included under two headers within the DESCRIPTION file. Dependencies under `Imports` are required for package installation whereas non-essential dependencies under `Suggests` are used for testing and building documentation such as vignettes. `knitr` is included under `Suggests` within the DESCRIPTION file. This seems like the proper place for it. (`Suggests` style dependencies are usually installed by setting the `dependencies` parameter to `TRUE` during installation. Otherwise the installer defaults to only installing the bare minimum of dependencies, those under `Imports`).

> Contributing to a repo does not just mean adding code. Your contribution guide should also mention where people should open issues and how to contribute new ideas/suggestions.

Addressed by #206 and #211

> The file data.R contains comments about the specific number of rows for datatype. Doesn't this depend on my input data? Or am I misunderstanding something.

`data.R` is describing specific sample data included with the package, rather than generic "accelerometer_data" or "gyroscope_data". When you load mhealthtools into the namespace these sample data are automatically imported as well.

> I am not sure if this is typical in the R-world, but a hosted documentation with browsable function doc-strings would be awesome (thinking ReadTheDocs in the Python world)

There is a popular package called [pkgdown](https://pkgdown.r-lib.org/) often used for this purpose. I'll look to see if this is something I can set up without too much additional effort.
