### 2020-02-27 c8ff0bc
additional installation instructions
