### 2020-02-27 Sage-Bionetworks/mhealthtools#209 @AKuederle
> > Provide more information about the vignettes in your README. I think they will be the starting point for most people who want to get started with the package.
>
> While addressing #206 I inserted a _Getting Help_ section after the high-level _Usage_ section. _Getting Help_ includes a sentence on what vignettes are provided. Do you think this is enough information? I don't want to be too specific in the README about the content of the vignettes since the vignettes are subject to change in the future.

That looks good to me!
