### 2020-02-27 Sage-Bionetworks/mhealthtools#209 @AKuederle
Thank you for clarifying these points. I think it was mostly my lack of experience in the R ecosystem that let to these questions.

Regrading the installation of the vignets:
Could you add the install commands required, to get everything working to the README? I think this would resolve all confusion.

Regarding pkgdown: I really think this is just a nice to have and for sure not a requirement!
