### 2020-02-27 Sage-Bionetworks/mhealthtools#209 @philerooski
I've merged #212 which makes changes to the example install commands. I tested the new instructions on the latest r-base docker image, and after installing devtools I was able to install mhealthtools+vignettes with the new command.

It turns out that building and publishing a docs website is incredibly easy with pkgdown + github pages (you can preview what it would look like here: https://philerooski.github.io/mHealthTools/). Though there is one small issue with referencing a figure from the paper using its relative path. This is something I would want to set up to work with Travis anyhow so I will add it to the to-do list and get around to it at a more convenient time.
