### 2020-03-19 cb78f07
first pass at related package comparison
