### 2020-03-24 858baa7
final touch ups to paper in response to feedback
