### 2020-03-24 Sage-Bionetworks/mhealthtools#210 @philerooski
Hi @AKuederle, I've added a section at the end which compares mhealthtools to similar packages in the digital health domain. If it's satisfactory to you we can close this issue.
