### 2020-03-27 #2106 @philerooski
Hi @trallard I believe things have wrapped up here.

I've made a tagged v0.2.0 release on the main repository and uploaded to Zenodo.

Zenodo: https://zenodo.org/record/3731707
Zenodo DOI: 10.5281/zenodo.3731707

There is one review check box left unchecked ("Performance") which I think should be checked. I did not make any performance claims in the paper.
